requirements.txt
########################################################## created at --> 2020-10-24_10-55-55 ##########################################################
aiohttp==3.6.3
appdirs==1.4.4
argcomplete==1.12.1
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
discord.py==1.5.1
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
idna==2.10
importlib-metadata==2.0.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
multidict==4.7.6
packaging==20.4
pipx==0.15.5.1
PyGithub==1.53
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
python-Levenshtein==0.12.0
pywin32==228
regex==2020.10.23
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
userpath==1.4.1
WMI==1.5.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.3.2
